+++ source_gen/source_gen/example/example.g.dart
@@ -1,5 +1,5 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
-// 2015-11-16T21:30:24.516Z
+// 2016-03-10T22:34:15.527866Z
 
 part of source_gen.example.example;