+++ source_gen/source_gen/lib/builder.dart
@@ -1,10 +1,5 @@
 // Copyright (c) 2015, the Dart project authors.  Please see the AUTHORS file
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
-library source_gen.build_file;
 
-import 'tool/build.dart' as build;
-
-main() async {
-  build.main();
-}
+export 'src/builder.dart';