+++ source_gen/source_gen/lib/source_gen.dart
@@ -1,7 +1,6 @@
 library source_gen;
 
-export 'src/build.dart';
-export 'src/generate.dart';
+export 'src/builder.dart';
 export 'src/results.dart';
 export 'src/generator.dart';
 export 'src/generator_for_annotation.dart';