+++ source_gen/source_gen/tool/build.dart
@@ -1,10 +1,12 @@
 // Copyright (c) 2015, the Dart project authors.  Please see the AUTHORS file
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
-library source_gen.build_file;
+library source_gen.tool.build;
 
-import 'tool/build.dart' as build;
+import 'package:build/build.dart';
+
+import 'phases.dart';
 
 main() async {
-  build.main();
+  build(phases);
 }